update way to take home dir and work with paths
Starting with Go 1.12, there is [os.UserHomeDir()](https://pkg.go.dev/os#UserHomeDir)
that should replace the `github.com/mitchellh/go-homedir` package
without any issues.

Safe cross-platform way to work with paths is
[path.Join](https://pkg.go.dev/path#Join) instead of string
concatenation. Windows doesn't care if there is a mixture of back and
forward slashes, but it's better to be on the safe side :)